+++ packages/pre-build/src/utils.js
@@ -134,3 +134,20 @@ exports.getNPMCommand = async () => {
 
   throw new Error(`Can not find these commands: ${candidate.join(', ')}`);
 };
+
+exports.execCallback = async (filePath, onSuccess) => {
+  const realOnSuccess = onSuccess || process.env.ON_SUCCESS_CALLBACK;
+  if (!realOnSuccess) {
+    return;
+  }
+  try {
+    await exports.exec(realOnSuccess, {
+      color: true,
+      env: {
+        HANDLED_FILE_PATH: filePath,
+      },
+    });
+  } catch (e) {
+    error(e);
+  }
+}